+++ tests/integration/cases/record_named_shape/Tcl@v8_6.tcl
@@ -1,0 +1,4 @@
+set my_data [list \
+    [dict create "id" 100 "description" "first task" "is_done" 0 "blocks" [list 102 103]] \
+    [dict create "id" 101 "description" "second task" "is_done" 1 "blocks" [list 100]] \
+]