+++ tests/integration/cases/record_named_shape/TypeScript_combined_declaration_style_var@v5.ts
@@ -1,0 +1,9 @@
+var my_data = [
+  {"id": 100, "description": "first task", "is_done": false, "blocks": [102, 103]},
+  {"id": 101, "description": "second task", "is_done": true, "blocks": [100]},
+];
+my_data = [
+  {"id": 100, "description": "first task", "is_done": false, "blocks": [102, 103]},
+  {"id": 101, "description": "second task", "is_done": true, "blocks": [100]},
+];
+export {};